Reverse a integer( leetcode problem) +  Duck No
In this program I have Merged two different problems together
First problem  is of leetcode to reverse a integer for which I've created a 
 seperated a function and 
second is to check it is a duck no or not.
Duck number is another special positive non-zero number that contains zero in it. 
The digit zero should not be presented at the starting of the number. 
Zero can be present at any of the positions except the beginning of the number.